requirements.txt
streamlit==1.30.0
python-pptx==1.0.2
openai==1.6.0
python-dotenv==1.0.0
requests==2.31.0
